requirements.txt
dnspython
pymongo
pyrogram
tgcrypto
python-dotenv
